requirements/test.txt
-r ./base.txt
pytest==5.1.2
pytest-django==3.5.1
